Fix circular import for DocumentContext
