+++ src/agentcore/state/contexts/documents/context.py
@@ -7,9 +7,8 @@ from agentcore.models import Document
 from agentcore.structures import ItemSequence
 from agentcore.structures.sequences import SequenceMixin
 
-from ..protocols import DocumentContext
 from .models import DocumentMatch, DocumentQuery
-from .protocols import DocumentStore
+from .protocols import DocumentContext, DocumentStore
 from .stores import InMemoryListStore
