+++ src/agentcore/state/contexts/protocols.py
@@ -50,16 +50,6 @@ class MessageContext(Sequence[ChatCompletionMessageParam], Protocol):
     def add(self, message: ChatCompletionMessageParam) -> None: ...
 
 
-class DocumentContext(Sequence[Document], Protocol):
-    # Existing simple API (kept for compatibility in this step)
-    def add(self, document: Document) -> None: ...
-
-    # New memory-like capabilities (store management and search)
-    def register_store(self, name: str, store: DocumentStore) -> None: ...
-    def store(self, name: str) -> DocumentStore: ...
-    def search(self, query: DocumentQuery) -> list[DocumentMatch]: ...
-
-
 class ToolContext(Mapping[str, Tool], Protocol): ...
